Packagify metarecord.admin
Convert metarecord.admin to a package so different model admin
views can be split into their own files.
