+++ metarecord/admin/__init__.py
@@ -1,0 +1,6 @@
+from .action import ActionAdmin
+from .attribute import AttributeAdmin, AttributeGroupAdmin
+from .classification import ClassificationAdmin
+from .function import FunctionAdmin
+from .phase import PhaseAdmin
+from .record import RecordAdmin
