+++ metarecord/admin/classification.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+from metarecord.models.classification import Classification
+
+
+@admin.register(Classification)
+class ClassificationAdmin(admin.ModelAdmin):
+    list_display = ('code', 'title', 'function_allowed')
+    ordering = ('code',)
